OTUS-198 - Alteração na referência dos validores para Double.

+++ survey-model/src/main/java/org/ccem/otus/survey/template/item/questions/fillingRules/validators/integer/In.java
@@ -14,8 +14,8 @@ public class In extends GenericValidator {
 		// Inner class
 		public class RangeInReference {
 			
-			public Integer initial;
-			public Integer end;
+			public Double initial;
+			public Double end;
 		}
 
 	}
